Move SubscriptionQuery class out of utils to dedicated file.
